Move UI into separate file
